+++ app/_components/molecules/ui/layouts/Header.tsx
@@ -1,5 +1,5 @@
 "use client";
-import Menu from "@/public/images/icons/icon_menu.svg";
+
 import Back from "@/public/images/icons/icon_arrow_left.svg";
 import logo from "@/public/images/logos/logo.korean.png";
 import Image from "next/image";
@@ -12,24 +12,7 @@ import { useRouter } from "next/navigation";
 import SlideMenu from "./SlideMenu";
 import { useState } from "react";
 
-export const HeaderDefault: React.FC = () => {
-  const [isSlideOpen, setIsSlideOpen] = useState<boolean>(false);
 
-  const onClickMenuHandler = () => {
-    setIsSlideOpen(true);
-  };
-
-  return (
-    <>
-      <HeaderWrapper left={false}>
-        <MenuButton onClick={() => onClickMenuHandler()}>
-          <Image src={Menu} alt={"메뉴 아이콘"} fill priority />
-        </MenuButton>
-      </HeaderWrapper>
-      {<SlideMenu isSlideOpen={isSlideOpen} setIsSlideOpen={setIsSlideOpen} />}
-    </>
-  );
-};
 
 export const HeaderBack: React.FC = () => {
   const router = useRouter();
